Fill out AggregateExecutorTests
- Changed the return type of the count function to Decimal, to account for numbers that are 2^31 and larger
- Fixed handling of alternate formats of date literals

+++ fhir-server/src/main/java/au/csiro/pathling/fhirpath/element/IntegerPath.java
@@ -61,9 +61,9 @@ public class IntegerPath extends ElementPath implements Materializable<Primitive
     }
     switch (getFhirType()) {
       case UNSIGNEDINT:
-        return Optional.of(new UnsignedIntType(row.getLong(columnNumber)));
+        return Optional.of(new UnsignedIntType(row.getInt(columnNumber)));
       case POSITIVEINT:
-        return Optional.of(new PositiveIntType(row.getLong(columnNumber)));
+        return Optional.of(new PositiveIntType(row.getInt(columnNumber)));
       default:
         return Optional.of(new IntegerType(row.getInt(columnNumber)));
     }
